- Import current version (v20100205-0000) of xrootd containing several fixes       in the client needed in particular by ATLAS; this includes the new SSL       module for authentication needed by Castor.     - Import patches #29438 and #29474 in krb5auth to fix building issues with current       linux distribution (e.g. kubuntu)
git-svn-id: http://root.cern.ch/svn/root/branches/v5-22-00-patches@32250 27541ba8-7e3a-0410-8455-c3a389f83636

+++ net/xrootd/src/xrootd/src/XrdOdc/XrdOdcFinder.cc
@@ -560,7 +560,8 @@ XrdOdcFinderTRG::XrdOdcFinderTRG(XrdSysLogger *lp, int whoami, int port)
    OLBp    = new XrdOucStream(&OdcEDest);
    Active  = 0;
    myPort  = port;
-   sprintf(buff, "login %c %d port %d\n",(isProxy ? 'P' : 'p'),getpid(),port);
+   sprintf(buff, "login %c %d port %d\n",(isProxy ? 'P' : 'p'),
+                 static_cast<int>(getpid()),port);
    Login = strdup(buff);
 }
